NEXUS-5912: Prevent extending components
As the "test" related components were doing
it, and also, those never got unregistered
from event bus.
